{"types":"C","runs":17}

+++ cons.java
@@ -1,15 +1,21 @@
 import acm.program.ConsoleProgram;
 
 public class cons extends ConsoleProgram {
-public void run(){
-	
-	double a = readDouble("enter first number : ");
-	double b = readDouble("enter second number : ");
-	double c = readDouble("enter third number : ");
-	println((a + b + c)/3);
-	
-	double d = (1 / a) + (1 / b) + (1 / c) ;
-	//System.out.print(1/a);
-	println( 1 / d );
-}
+	public void run(){
+		
+		int a = readInt("enter first number : ");
+		int b = readInt("enter second number : ");
+		int c = readInt("enter third number : ");
+		println((a + b + c)/3);
+		
+		double d = (int)(1 / a) + (1 / b) + (1 / c) ;
+		//System.out.print(1/a);
+		println( 1 / d );
+		
+		
+		
+		
+		
+		
+	}
 }
